prototype/requirements.txt
streamlit>=1.32.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.18.0
openai>=1.0.0
google-generativeai>=0.5.0
